Creates search bar to filter by min and max price as well as name in products tab
